pyproject.toml
[project]
name = "mystoreke"
version = "0.1.0"
description = "Python backend for mystoreke.com (FastAPI + Supabase)."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "python-dotenv>=1.0",
  "jinja2>=3.1",
  "supabase>=2.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4",
]

[tool.uvicorn]
factory = false
